Bad bug fixed
Before the fix the situation was as follows. Let's say we selected
a function cfg for which can be painted (function 1). Then we miminized
Control-flow graph window and selected a functin cfg for which cannot
be painted because objdump output contains too many empty lines
(function 2). After that we maximized Control-flow graph window and
got segmentation fault. It was so because the scene still contained
canvas parts of function 2 but corresponding cfg parts had already
been destroyed.

Clearing the scene inside ControlFlowGraph::reset fixed this bug.